Update Jaeger dependency to 0.27.0
This commit updates the io.jaegertracing dependencies to 0.27.0 from
0.27.0-RC1. The update exposed an issue with the shaded thrift92 JAR,
namely that it bundles SLF4J and Apache HTTP dependencies that don't
show up in the project dependency map. This causes version conflicts
that are dependent on the order of the classpath.

So, we stop using the shaded jar and exclude unwanted dependencies
brought in by Jaeger components.